understanding async requests

+++ controllers/admin.js
@@ -173,21 +173,39 @@ exports.getProducts = (req, res, next) => {
     });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-  const prodId = req.body.productId;
+// exports.postDeleteProduct = (req, res, next) => {
+//   const prodId = req.body.productId;
+//   Product.findById(prodId)
+//     .then((product) => {
+//       fileHelper.deleteFile(product.imageUrl);
+//     })
+//     .catch((err) => net(err));
+
+//   Product.deleteOne({ _id: prodId, userId: req.user._id })
+//     .then(() => {
+//       res.redirect('/admin/products');
+//     })
+//     .catch((err) => {
+//       const error = new Error(err);
+//       error.httpStatusCode = 500;
+//       return next(error);
+//     });
+// };
+
+exports.deleteProduct = (req, res, next) => {
+  const prodId = req.params.productId;
   Product.findById(prodId)
     .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
       fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
     })
-    .catch((err) => net(err));
-
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
     .then(() => {
-      res.redirect('/admin/products');
+      res.status(200).json({ message: 'Success!' });
     })
     .catch((err) => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
+      res.status(500).json({ message: 'Deleted product failed.' });
     });
 };
